kcons: handle depfiles in makefile
An extra rule called .depend is added to Makefie.i386. It will generate
a depfile called .depend with additional rules that can be used by the
makefile to infer changes to the header files. These would allow to
recompile an object file if it one of the header files the source file
includes is modified.

The change has been tested with gmake and bmake to assert multiple
conditions, including compiling the kernel without generating a depfile
first. gmake seems to always recompile the depfile. bmake will not
unless told so, but at least things do not fail. However, making changes
to a header file without a valid depfile will cause the object file to
not be recompiled.