releng: add buildno.sh script
buildno.sh will generate a valid build number for compilations using the
Git repository as source. It will count the build number as the total
number of commits made to the trunk branch since the beginning of the
project. I do not know exactly what is the beginning of the project
because the source code has been rewritten a lot during these years, but
it's something.

If the script is called outside of a Git repository (such as a tarball
or zipball snapshot), it will silently fail.