doc: update doxygen toolchain
I'm not listening a lot into Doxygen and I should listen to it more. It
turns out that things are not being documented as expected. This commit
updates Doxygen to the latest version, and also makes some changes:

* Enhances the way Doxygen treats the briefs: autobrief, do not repeat.
* Tells Doxygen that this is a C project - so it will call things data
  structures rather than classes.

Most of the remaining parts of the Doxyfile diff are changes to the
comments. I've just executed doxygen -u to update the file, so these
changes comes from upstream.