vfs: handle fs_resolve requests to the root
When fs_resolve receives a request to access the vfs_node at the root of
a mountpoint (such as DEV:/ or SYS:/), it will fail to return the
vfs_node_t entry that points to the mountpoint root.

This commit addresses an empty path after the volume has been resolved,
so that if it happens, it returns the root instead.