vfs: wrap node_t operations in vfs_ops_t
Because more hooks will be added in the future, this commit will add a
new data structure to hold all the hooks called vfs_ops_t, which is a
hook table that contains all the function hooks used by a specific
file system implementation.

At the moment, the vn_ops field has been added to the vfs_node_t data
structure, and it is shared by all the VFS nodes part of the same
virtual file system. However, this behaviour is subject to change in the
future, as more data structures for the VFS subsystem are added.