vfs: linker set for the vfs drivers
In this commit, every existing FS implementation will receive a new
struct of type vfs_filesys_t. This data structure represents a file
system driver. In future commits, it will gain more hooks such as the
hooks used to mount volumes.

At the moment, this commit introduces a linker set similar to the one
added for devices. The FS_DESCRIPTOR() macro will create a pointer to a
specific vfs_filesys_t node in a separate section of the ELF image.
During the system startup process, the section is traversed and every
pointer is dereferenced, initialised, and added to a file system menu.