vfs: add vfs_volume_t data structure
The volume represents an instance mounted in a file system. For
instance, the driver for a FAT32 file system may have multiple volumes,
one per hard drive. Some file systems will only have one volume, like
DEVFS or ROOTFS.

Once a driver knows about its volumes, it is possible to change the
vfs_mount function so that it accepts a driver, a parameter, and the
volume name, and it handles mounting multiple volumes of the same
driver.

vfs_driver_t has gained a new field called fsd_mount. This is a hook
that tells the driver that a volume is being mounted. Drivers can use
this to reject the mount, or to setup additional data structures.

The rest of this commit is basically fixing bugs after changing the data
structures:

* Replace mountpoint_t with vfs_volume_t, and change all the lists using
  this data structure.
* rootfs_create_entry has been replaced with a different list (because
  there is no mountpoint_t anymore).