vfs: handle fields present in vfs_node_t
After this commit, an vfs_node_t will not be connected to the vfs_ops_t
data structure. The VFS Operations Table is shared between all the nodes
of the underlying VFS Driver, so there is no need to make a node point
to the table, as long as the node points to the belonging driver.

vfs_node_t has gained a field for pointing to the vfs_volume_t instance
owning the node, which is required also for understanding shared
payloads when the VFS Driver allows mounting multiple volumes of the
same VFS Driver at the same time.

To make code in the fsops module easier to follow, a NODE_OPS macro has
been added, to do the traversing to get the VFS Operations Table for a
specific node, as long as the path is not broken. the fsops module has
been refactored to use this macro already.