stdkern: implement strchr and strrchr functions
strchr is a function that will return a pointer to the first ocurrence
of the given character in a string. strrchr is a function that returns a
pointer to the last ocurrence of the given character in the string.

Related to these functions, the strchrnul function is not implemented.
This is becaus strchrnul is a GNU extension and therefore I doubt I am
going to require this inside the kernel, although the userland libc will
have to provide it probably.