kcons: generate a config.h file
This commit will add two new directives to the configuration language
used by the kcons tool, called define and nodefine.

The define directive is used to annotate a macro definition. These are
collected when the compile directory is being generated, translated into
C preprocessor macros, and then saved into a file called config.h, that
will be generated in the compilation directory. This file can be
imported by the kernel files in order to have conditional definitions
set by the profile.

The nodefine directive allows to remove a previously set directive.