i386: split boot in locore + multiboot
It's not like if I wanted to drop support for Multiboot anytime soon (I
hope), but this commit will make use of support for defines previously
added to kcons in 7044235 (kcons: generate a config.h file, 2022-01-11)
in order to generate a macro that confirms whether NativeOS is being
built with support for GNU Multiboot (MULTIBOOT).

The routines related to Multiboot have been removed from boot.S and
placed into a different file called multiboot.S. The lonely definition
of a multiboot_info_t that used to be in multiboot.c has also been moved
into multiboot.S and is now defined using standard Assembly code.

boot.S has been renamed to locore.S and it is now placed in the new
folder where i386 code should be placed. I'm taking this to also place
in kernel/i386 the ldscript (kernel.ld). References in the compile files
have been updated.