i386: cleanup multiboot.h data structures
In this commit, multiboot.h is moved to the new destination in
kernel/i386. Files that reference this import have had their definitions
updated too.

However, the point of this commit was to do some cleanup to the
multiboot.h file. Because of the nature of the changes, Git has lost
track of the rename, and it treats the old and the new file as something
separate, but the list of changes include:

* I've removed __attribute__((packed)) from the data structures,
  considering that it is not necessary because the compiler will pack
  properly these data structures.

* I've removed some data structures that are not going to be used, such
  as the a.out information. (The kernel will boot as an ELF application,
  so I do not need this).

* Cleanup the documentation and fixing the comments. Many of these were
  written some years ago when I did not even know what did they mean. (I
  chuckled a bit when I read that I did not expect to have to read the
  ELF section table...)