build: remove stash files that will not be used
Some of the files in the arch/ and contrib/ directories is not going to
be used because it is not in an usable state. For this cases, the best
solution would be to simply drop these files. If they are ever needed,
they can be checked from the Git repository in order to be rewritten as
something that worked better.

printk did not handle most placeholders, plus it relies on an VGA driver
that is gone. The VBE driver only contains the most basic code, and it
might be better to just integrate a new VBE driver. The paging.h file is
not going to be used for the VMM system.