vtcon: virtual multiple consoles
An upgrade has been made to the virtual console driver to support
multiple consoles. After applying this commit, it will be possible to
switch the framebuffer screen to one of the eight supported contexts, by
using the F1 to F8 keys in the keyboard.

To indicate the current virtual console in use, a context bar has been
added at the bottom of the screen that presents the context number, as
well as some other "useful" information for the preview builds that I'm
generating.