tarfs: initial version of the TARFS driver
TARFS is a file system driver capable of reading buffers containing TAR
files. TARFS is mainly used for designing ramdisks that can be included
in the CD-ROM or provided by the bootloader when the operating system
starts, so that a "live CD" environment can be prepared.

The TARFS driver only supports read only file systems, therefore the
driver does not have definitions for write functions. It is only
possible to traverse a memory buffer that contains a TAR file, and read
the files.