rtclock: do not use interrupts for compatibility
It seems that some machines do not like to run interrupts to know when
the time passes. Therefore, in this commit time is polled actively.
Every time you read from DEV:/CLOCK, the time will be fetched.

(This should probably be a system call, but there are no system calls in
the kernel yet.)